chore: move cities to exteral file
