+++ src/helpers/extractCity.js
@@ -1,47 +1,5 @@
 const fuzz = require("fuzzball");
-
-const cities = [
-  "Aden",
-  "Alexandria",
-  "Amsterdam",
-  "Athens",
-  "Basrah",
-  "Boston",
-  "Brunei",
-  "Buenos Aires",
-  "Calicut",
-  "Kolkata",
-  "Cape Town",
-  "Cayenne",
-  "Ceylon",
-  "Copenhagen",
-  "Darwin",
-  "Edo",
-  "Hamburg",
-  "Hangzhou",
-  "Istanbul",
-  "Jamaica",
-  "Las Palmas",
-  "Lisbon",
-  "London",
-  "Luanda",
-  "Malacca",
-  "Manila",
-  "Marseille",
-  "Mozambique",
-  "Nantes",
-  "Nassau",
-  "Panama City",
-  "Pinjarra",
-  "Quanzhou",
-  "Rio De Janeiro",
-  "Santo Domingo",
-  "Seville",
-  "St. George's",
-  "Stockholm",
-  "Tunis",
-  "Venice",
-];
+const cities = require("../data/cities.js");
 
 function matchCityName(input) {
   let match = null;
